refactor(tui): extract seat geometry to pixel_painter/seat.rs
Move the cohesive seat-orientation unit out of pixel_painter/mod.rs into
seat.rs: SeatView (the single source of truth for how a waypoint occupant
faces — sprite/flip/sit-down-glide/z-key), seat_sprite, settle_seat_view,
DESK_SEAT_Z_OFF, and the shared recolor-blit paint_character_at.

mod.rs 1350 lines (2736 -> 1350, -51% across this reorg arc). seat.rs uses
`use super::*;` for its palette/cache/anchors deps (same as the test
sibling); mod.rs re-imports the four symbols it calls; drawable.rs reaches
paint_character_at via the unchanged `super::` path; the test module reaches
DESK_SEAT_Z_OFF via `super::seat::`. New mod decls grouped into the existing
alphabetical block. tui/CLAUDE.md updated. 823 tests pass.
