+++ crates/pixtuoid/src/tui/CLAUDE.md
@@ -57,6 +57,7 @@ tui/
                     resolved ONCE per frame [8 bundled / 10 robot] — so a non-8-wide pack centers correctly),
                     furniture.rs (coffee table, area rug, side table, pantry table/chair),
                     glass.rs (frosted-glass room-divider walls: consts + stitch + paint_glass_wall_h/v),
+                    seat.rs (SeatView orientation single-source + seat_sprite + settle_seat_view + paint_character_at),
                     tests.rs (sibling unit suite, extracted from mod.rs)
 ```
